feat: CurrentLocationView 생성, WeatherView에서 장소만 제공
