+++ WeatherApp/View/CurrentLocationView.swift
@@ -1,0 +1,24 @@
+//
+//  LocationView.swift
+//  WeatherApp
+//
+//  Created by 정현 on 4/12/24.
+//
+
+import SwiftUI
+import WeatherKit
+
+struct CurrentLocationView: View {
+    @State var weather : Weather
+    @State var viewModel : ContentViewViewModel
+    var body: some View {
+        VStack(spacing : 5) {
+            Text(viewModel.weatherManager.locationManager.city)
+                .font(.title)
+        }
+    }
+}
+
+//#Preview {
+//    LocationView()
+//}
