--- WeatherApp/View/LocationView.swift
@@ -1,18 +1,0 @@
-//
-//  LocationView.swift
-//  WeatherApp
-//
-//  Created by 정현 on 4/12/24.
-//
-
-import SwiftUI
-
-struct LocationView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    LocationView()
-}
