--- WeatherApp/View/TodayWeatherView.swift
@@ -1,18 +1,0 @@
-//
-//  TodayWeatherView.swift
-//  WeatherApp
-//
-//  Created by 정현 on 4/12/24.
-//
-
-import SwiftUI
-
-struct TodayWeatherView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    TodayWeatherView()
-}
